+++ lprint-sii.h
@@ -1,0 +1,20 @@
+//
+// Seiko Instruments, Inc. driver table for LPrint, a Label Printer Application
+//
+// Copyright © 2023 by Michael R Sweet.
+//
+// Licensed under Apache License v2.0.  See the file "LICENSE" for more
+// information.
+//
+
+{ "sii_slp100_203dpi", "SII SLP100", NULL, NULL },
+{ "sii_slp200_203dpi", "SII SLP200", NULL, NULL },
+{ "sii_slp240_203dpi", "SII SLP240", NULL, NULL },
+{ "sii_slp410_203dpi", "SII SLP410", NULL, NULL },
+{ "sii_slp420_203dpi", "SII SLP420", NULL, NULL },
+{ "sii_slp430_203dpi", "SII SLP430", NULL, NULL },
+{ "sii_slp440_300dpi", "SII SLP440", NULL, NULL },
+{ "sii_slp450_300dpi", "SII SLP450", NULL, NULL },
+{ "sii_slp620_203dpi", "SII SLP620", NULL, NULL },
+{ "sii_slp650_300dpi", "SII SLP650", NULL, NULL },
+{ "sii_slp650se_300dpi", "SII SLP650SE", NULL, NULL },