+++ lprint.c
@@ -53,7 +53,7 @@ static pappl_pr_driver_t	lprint_drivers[] =
 #include "lprint-brother.h"
 #include "lprint-dymo.h"
 #include "lprint-epl2.h"
-#include "lprint-seiko.h"
+#include "lprint-sii.h"
 #include "lprint-tspl.h"
 #include "lprint-zpl.h"
 };
@@ -185,6 +185,7 @@ driver_cb(
     ipp_t                  **attrs,	// O - Pointer to driver attributes
     void                   *cbdata)	// I - Callback data (not used)
 {
+  bool	ret = false;			// Return value
   int	i;				// Looping var
 
 
@@ -239,14 +240,39 @@ driver_cb(
   data->testpage_cb = lprintTestPageCB;
 
   // Use the corresponding sub-driver callback to set things up...
-  if (!strncmp(driver_name, "dymo_", 5))
-    return (lprintDYMO(system, driver_name, device_uri, device_id, data, attrs, cbdata));
+  if (!strncmp(driver_name, "brother_", 8))
+    ret = lprintBrother(system, driver_name, device_uri, device_id, data, attrs, cbdata);
+  else if (!strncmp(driver_name, "dymo_", 5))
+    ret = lprintDYMO(system, driver_name, device_uri, device_id, data, attrs, cbdata);
   else if (!strncmp(driver_name, "epl2_", 5))
-    return (lprintEPL2(system, driver_name, device_uri, device_id, data, attrs, cbdata));
+    ret = lprintEPL2(system, driver_name, device_uri, device_id, data, attrs, cbdata);
+  else if (!strncmp(driver_name, "sii_", 4))
+    ret = lprintSII(system, driver_name, device_uri, device_id, data, attrs, cbdata);
+  else if (!strncmp(driver_name, "tspl_", 5))
+    ret = lprintTSPL(system, driver_name, device_uri, device_id, data, attrs, cbdata);
   else if (!strncmp(driver_name, "zpl_", 4))
-    return (lprintZPL(system, driver_name, device_uri, device_id, data, attrs, cbdata));
-  else
-    return (false);
+    ret = lprintZPL(system, driver_name, device_uri, device_id, data, attrs, cbdata);
+
+  // Update the ready media...
+  for (i = 0; i < data->num_source; i ++)
+  {
+    pwg_media_t *pwg = pwgMediaForPWG(data->media_ready[i].size_name);
+
+    data->media_ready[i].bottom_margin = data->bottom_top;
+    data->media_ready[i].left_margin   = data->left_right;
+    data->media_ready[i].right_margin  = data->left_right;
+    data->media_ready[i].size_width    = pwg->width;
+    data->media_ready[i].size_length   = pwg->length;
+    data->media_ready[i].top_margin    = data->bottom_top;
+    papplCopyString(data->media_ready[i].source, data->source[i], sizeof(data->media_ready[i].source));
+    papplCopyString(data->media_ready[i].type, data->type[0], sizeof(data->media_ready[i].type));
+  }
+
+  // By default use media from the main source...
+  data->media_default = data->media_ready[0];
+
+  // Return...
+  return (ret);
 }