+++ lprint.h
@@ -107,6 +107,8 @@ typedef ipp_copycb_t ipp_copy_cb_t;
 #  define LPRINT_TESTPAGE_HEADER	"T*E*S*T*P*A*G*E*"
 
 #  define LPRINT_EPL2_MIMETYPE		"application/vnd.eltron-epl"
+#  define LPRINT_SLP_MIMETYPE		"application/vnd.sii-slp"
+#  define LPRINT_TSPL_MIMETYPE		"application/vnd.tsc-tspl"
 #  define LPRINT_ZPL_MIMETYPE		"application/vnd.zebra-zpl"
 
 
@@ -142,7 +144,7 @@ extern bool	lprintDitherLine(lprint_dither_t *dither, unsigned y, const unsigned
 extern bool	lprintBrother(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
 extern bool	lprintDYMO(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
 extern bool	lprintEPL2(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
-extern bool	lprintSEIKO(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
+extern bool	lprintSII(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
 extern bool	lprintTestFilterCB(pappl_job_t *job, pappl_device_t *device, void *data);
 extern const char *lprintTestPageCB(pappl_printer_t *printer, char *buffer, size_t bufsize);
 extern bool	lprintTSPL(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);